Unify Tequilapi server & client DTOs for healthcheck
